Example dataset from a label-free DIA, a group comparison study of S.Pyogenes.

This example dataset was obtained from a group comparison study of S. Pyogenes. Two conditions, S. Pyogenes with 0% and 10% of human plasma added (denoted Strep 0% and Strep 10%), were profiled in two replicates, in the label-free mode, with a SWATH-MS-enabled AB SCIEX TripleTOF 5600 System. The identification and quantification of spectral peaks was assisted by a spectral library, and was performed using OpenSWATH software (http: //proteomics.ethz.ch/openswath.html). For reasons of space, the example dataset only contains two proteins from this study. Protein FabG shows strong evidence of differential abundance, while protein Probable RNA helicase exp9 only shows moderate evidence of dif- ferential abundance between conditions.

DIARawData

Format

data.frame

Details

The raw data (input data for MSstats) is required to contain variable of ProteinName, PeptideSequence, PrecursorCharge, FragmentIon, ProductCharge, IsotopeLabelType, Condition, BioReplicate, Run, Intensity. The variable names should be fixed.

If the information of one or more columns is not available for the original raw data, please retain the column variables and type in fixed value. For example, the original raw data does not contain the information of PrecursorCharge and ProductCharge, we retain the column PrecursorCharge and ProductCharge and then type in NA for all transitions in RawData.

Variable Intensity is required to be original signal without any log transformation and can be specified as the peak of height or the peak of area under curve.

Value

data.frame with the required format of MSstats.

Examples

head(DIARawData)

#>          ProteinName  PeptideSequence PrecursorCharge FragmentIon ProductCharge
#> 1 RNA  helicase exp9 ASPIQEMTIPLALEGK               2          b9         77414
#> 2 RNA  helicase exp9 ASPIQEMTIPLALEGK               2         y10         77411
#> 3 RNA  helicase exp9 ASPIQEMTIPLALEGK               2         y11         77412
#> 4 RNA  helicase exp9 ASPIQEMTIPLALEGK               2         y14         77410
#> 5 RNA  helicase exp9 ASPIQEMTIPLALEGK               2          y7         77409
#> 6 RNA  helicase exp9 ASPIQEMTIPLALEGK               2          y9         77413
#>   IsotopeLabelType Condition BioReplicate Run Intensity
#> 1                L  Strep 0%            1   1      9747
#> 2                L  Strep 0%            1   1      1272
#> 3                L  Strep 0%            1   1      1295
#> 4                L  Strep 0%            1   1      2332
#> 5                L  Strep 0%            1   1      8178
#> 6                L  Strep 0%            1   1      1737